Made exceptions handling expansive to catch everything
Made exceptions handling expansive to catch everything so don't keep having to add specific exceptions.   This is ok as the exception handling causes a retry to be initiated with a timeout  and in vehicles.py also a retry count limit.  The Wafbypass  on the other hand is retired forever every 10 minutes but this should be fine for the operation for now since we always need the bypass to work to do anything.