Added Dealer Zip, City, Lat, Long, Distance to  searchForVehicles.py and distance formula to the vehiicles.py csv creatinon
Updated searchForVehicles.py to include new fields for display and compare  (exclude the CenterLat, CenterLong, and DistanceFromCenter fields since these are more for the user to modify when open the .csv manually.    In the  vehicles.py creation of the .csv added the CenterLat, CenterLong, and DistanceFromCenter  columns and a formula for calculating the distance from center to the Dealer lat Long location. the user can then pull down in the file when opened with excel.