Updates so that comparing NAN values in detailsAreTheSame yields true, plus added test files
In detailsAreTheSame  and getNamesOfModifiedFieldsIntoString ,  if the old value was nan or None and the new value was nan or None,  we treat comparing the old and new as True.  Normally if you compare using == a nan with a nan it yields false
Added new test files to test this out and removed old ones.