Added new CenterLatLong to config items and distance calculations to SearchForVehicles
So can now specify the CenterLatLong to show distances from and can use this in the user match criteria  now to limit radius to look for a vehicle.   This is useful compared to the environment variables for zipcode and radius  as you can run several different config files on the same local inventory data file without having to get the inventory data each time from the toyota website.   Also you are not limited to 10,000 vehicles like the zipcode environment variable search is.