Added testing of the config items  outputResultsMethod, alsoNotifyOnOnlyRemovals, and showRemovalsWhenOutputStatusIsAll
The testing was for the config item outputResultsMethod == outputAllSearchResultsOnChange and the expected behavior on that due to various settings of  config items alsoNotifyOnOnlyRemovals, and showRemovalsWhenOutputStatusIsAll
The toplevel test batch file for this is ConfigItemOutputResultsMethodTest.bat.   Currently that file only tests the outputResultsMethod == outputAllSearchResultsOnChange selection but others will be added in the future.