When inventory removed due to match filtering show modified field names
When the old and new unfiltered inventory has the same VIN but match filtering does not have that VIN anymore,  then when removed is to be shown , we show the new unfiltered VIN data as the removed data and if there are differences between that and the old matches filtering,  the  Names of Modified fields shows the diff between that and the old matches VIN data.  This is done with the removed prefix.     This is done so that the user can see that the VIN was removed not because it was no longer in the unfiltered inventory, but because it failed to pass the filtering match criteria.