Added automatically creating Sold inventory files, and a first time VIN added column
Added automatically creating Sold inventory files which are inventory that no longer appears in the gotten inventory from the website (i.e. was there then disappeared)
Also  added a FirstAddedDate column to the spreadsheet that is the first time that VIN appeared in the inventory to all people to easily when new VINs are added.  Modified upload-files to account for the new file names.