Added optional Dealers Adders file for dealers that are not returned by the website
For at least 1 zip code the website did not have that dealer listed under that.   So have and Adders json file that those that the website does not have can be manually placed there to account for this.  This file when added as a parameter on the dealers.py command line will update the file if those dealers in the Adders are not in the file alread or returned by the website for that run.