Added Change History feature. See readme.txt for more detail
The vehiclesTestWithChangeHistoryIndividualColumns.py is an archive of an attempt at showing each column change by adding 2 columns for each column.  This was abandoned because the resulting spreadsheet was to cluttered and it was difficult to identify changes.   Instead all changes are shown as text in a single List of Changes  Column.