Excel distance formula now created  based on where the input columsn are
Cell references for these input columns (latitudes and longitudes) now  automatically calculated and used in the excel  distance formula so no longer assuming the same fixed position for these input columns.  This supports csv files where those input columns are in different column positions than the normal csv files (such as is the case when it is a Change History file where some columns are added at the start.