Added another apparently alias dealer name
TOYOTA OF WARREN  with dealerCd of 34143  is the same as Golling Toyota of Warren MI (dealerCd 21121).
Did this so that older entries in Sold files will not give a missing dealer state error message each time the inventory is run as this is inventory that no longer appears and thus is not automatically corrected.